add: cascade deleting in tables, changes to make it work

+++ src/main/java/com/github/resalner/javapractice/service/ProductServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import com.github.resalner.javapractice.exception.EntityNotFoundException;
 import com.github.resalner.javapractice.model.Category;
 import com.github.resalner.javapractice.model.Product;
+import com.github.resalner.javapractice.repository.CategoryRepository;
 import com.github.resalner.javapractice.repository.ProductRepository;
 
 import jakarta.persistence.Column;
@@ -20,7 +21,8 @@ import java.util.Objects;
 public class ProductServiceImpl implements ProductService {
 
 	private final ProductRepository productRepository;
-
+	private final CategoryRepository categoryRepository;
+	
 	@Override
 	public List<Product> getProducts() {
 		return productRepository.findAll();
@@ -28,8 +30,13 @@ public class ProductServiceImpl implements ProductService {
 
 	@Override
 	public Product saveProduct(Product product) {
-		product = productRepository.save(product);
-		return product;
+		Category category = product.getCategory();
+        if (category != null && category.getId() != null) {
+            categoryRepository.findById(category.getId())
+                .orElseThrow(() -> new EntityNotFoundException("Категория с id = " + category.getId() + " не найдена"));
+        }
+        
+        return productRepository.save(product);
 	}
 
 	@Override
@@ -40,6 +47,10 @@ public class ProductServiceImpl implements ProductService {
 
 	@Override
 	public void deleteProduct(long id) {
+		if (!productRepository.existsById(id)) {
+	        throw new EntityNotFoundException("Не найден продукт с id = " + id);
+	    }
+		
 		productRepository.deleteById(id);
 	}
